chore: rename dirs to better reflect contents
