+++ src/ui/scripts/components/footer.js
