+++ src/ui/scripts/components/header.js
