+++ src/ui/scripts/components/workers.js
