+++ src/ui/scripts/models/work-item.js
