+++ src/ui/scripts/pages/work-open.js
