+++ webpack.dev.js
@@ -5,7 +5,7 @@ module.exports = merge(common, {
   mode: 'development',
   devtool: 'inline-source-map',
   devServer: {
-    contentBase: './dist-app',
+    contentBase: './dist-ui',
     writeToDisk: true,
     proxy: {
       '/.netlify/functions': {
